requirements.txt
streamlit
Pillow
face_recognition
opencv-python
numpy
